Add basic SOFT sample support (incomplete)
Working on adding support for NCBI GEO SOFT files that contain sample
data (until this point, GEXPRO can only handle curated datasets). This
involves making the parser first recognize that it has encountered a
Sample block, and then parsing that block until its end, before
reverting to the base state in preparation for the next block (which
may be another sample). This is only partially complete -- I'm not
sure how GEXPRO will handle alternating between adjacent sample blocks
and other blocks.

ADDITIONALLY, GEXPRO cannot yet take the read samples and build a data
matrix from them. This is the next stage. However, in order to do
this, I need to make sure it is pulling the correct columns from the
input file and that it can always convert the VALUE field to a
float (e.g., via stof()).

Now that I have returned to this project after a long time away, I can
tackle the issues of consistency. It helps that I took the time to
read the documentation of the SOFT file format in the meantime :)