Add simple keras autoencoder with TensorBoard support.
The autoencoder trains on the normalized data! Obviously, this doesn't mean that
the AE is actually learning anything meaningful, but it means that the training
task is computationally tractable. The pure TF version of the autoencoder is not
adequate for training---need to compare it to the Keras version's compiled graph
and see what is making the difference.