Fix handling of multiple GEXPROs.
mkgexpro would segfault with EXC_BAD_ACCESS upon trying to flush gexpros to
files. It turns out that the data was being deallocated prematurely, since we
were only keeping pointers to the gexpros in MultiGexpro. The fix for this is to
copy the entire gexpro objects into MultiGexpro.

Moving forward I will make sure that---after loading them into an instance of
MultiGexpro---the individual gexpro objects are passed only by reference, and
will ensure that no extraneous copies of the gexpros are floating around
elsewhere.